Subset downloaded data

hpaSubset() subsets data by gene name, tissue, cell type, cancer and/or cell line. The input is the list object generated by hpaDownload() or as the output of another hpaSubset(). Use hpaListParam() to see the list of available parameters for a specific list object. This is a convenient wrapper for `lapply/filter` and works on any table which contain 'gene', 'tissue', 'cell_type', 'cancer', and 'cell_line' columns.

hpaListParam() list available variables in downloaded data that can be used as parameters to subset the data via hpaSubset(). This function work with the data object generated by hpaDownload() or a previous call of hpaSubset(). This is a convenient wrapper for `lapply/unique` and works on any table which contain 'tissue', 'cell_type', 'cancer', and 'cell_line' columns.

Usage

hpaSubset(
  data = NULL,
  targetGene = NULL,
  targetTissue = NULL,
  targetCellType = NULL,
  targetCancer = NULL,
  targetCellLine = NULL
)

hpaListParam(data = NULL)

Arguments

data

Input the list object generated by hpaDownload() or hpaSubset()

targetGene

Vector of strings of HGNC gene symbols. It will be used to subset every dataset in the list object. You can also mix HGNC gene symbols and ensemnbl ids (start with ENSG) and they will be converted to HGNC gene symbols.

targetTissue

Vector of strings of normal tissues. Will be used to subset the normal_tissue and rna_tissue dataset.

targetCellType

Vector of strings of normal cell types. Will be used to subset the normal_tissue dataset.

targetCancer

Vector of strings of cancer types. Will be used to subset the pathology dataset.

targetCellLine

Vector of strings of cell lines. Will be used to subset the rna_cell_line dataset.

Value

hpaSubset will return a list of tibbles as the result of subsetting, depending on the input data.

The output of hpaListParam() is a list of vectors containing all subset parameter for the downloaded data.

See also

Examples

  downloadedData <- hpaDownload(downloadList='histology', version='example')
#> Only the followings are example/built-in datasets: 
#>  - Normal tissue 
#>  - Pathology 
#>  - Subcellular location 
#> Other datasets will not be loaded
  geneList <- c('TP53', 'EGFR')
  tissueList <- c('breast', 'cerebellum', 'skin 1')
  cancerList <- c('breast cancer', 'glioma', 'melanoma')

  subsetData <- hpaSubset(data=downloadedData,
                          targetGene=geneList,
                          targetTissue=tissueList,
                          targetCancer=cancerList)

  downloadedData <- hpaDownload(downloadList='histology', version='example')
#> Only the followings are example/built-in datasets: 
#>  - Normal tissue 
#>  - Pathology 
#>  - Subcellular location 
#> Other datasets will not be loaded
  params <- hpaListParam(data=downloadedData)
  params$normal_tissue
#> $tissue
#>  [1] "Adipose tissue"       "Adrenal gland"        "Appendix"            
#>  [4] "Bone marrow"          "Breast"               "Bronchus"            
#>  [7] "Caudate"              "Cerebellum"           "Cerebral cortex"     
#> [10] "Cervix"               "Colon"                "Duodenum"            
#> [13] "Endometrium"          "Epididymis"           "Esophagus"           
#> [16] "Fallopian tube"       "Gallbladder"          "Heart muscle"        
#> [19] "Hippocampus"          "Kidney"               "Liver"               
#> [22] "Lung"                 "Lymph node"           "Nasopharynx"         
#> [25] "Oral mucosa"          "Ovary"                "Pancreas"            
#> [28] "Parathyroid gland"    "Placenta"             "Prostate"            
#> [31] "Rectum"               "Salivary gland"       "Seminal vesicle"     
#> [34] "Skeletal muscle"      "Skin"                 "Small intestine"     
#> [37] "Smooth muscle"        "Soft tissue"          "Spleen"              
#> [40] "Stomach"              "Testis"               "Thyroid gland"       
#> [43] "Tonsil"               "Urinary bladder"      "Vagina"              
#> [46] "N/A"                  "Hypothalamus"         "Choroid plexus"      
#> [49] "Efferent ducts"       "Hair"                 "Retina"              
#> [52] "Lactating breast"     "Thymus"               "Cartilage"           
#> [55] "Eye"                  "Pituitary gland"      "Dorsal raphe"        
#> [58] "Brown adipose tissue" "Substantia nigra"     "Sole of foot"        
#> 
#> $cell_type
#>   [1] "adipocytes"                                
#>   [2] "glandular cells"                           
#>   [3] "lymphoid tissue"                           
#>   [4] "hematopoietic cells"                       
#>   [5] "myoepithelial cells"                       
#>   [6] "respiratory epithelial cells"              
#>   [7] "glial cells"                               
#>   [8] "neuronal cells"                            
#>   [9] "cells in granular layer"                   
#>  [10] "cells in molecular layer"                  
#>  [11] "Purkinje cells"                            
#>  [12] "endothelial cells"                         
#>  [13] "neuropil"                                  
#>  [14] "squamous epithelial cells"                 
#>  [15] "peripheral nerve/ganglion"                 
#>  [16] "cells in endometrial stroma"               
#>  [17] "cardiomyocytes"                            
#>  [18] "cells in glomeruli"                        
#>  [19] "cells in tubules"                          
#>  [20] "cholangiocytes"                            
#>  [21] "hepatocytes"                               
#>  [22] "alveolar cells"                            
#>  [23] "macrophages"                               
#>  [24] "germinal center cells"                     
#>  [25] "non-germinal center cells"                 
#>  [26] "ovarian stroma cells"                      
#>  [27] "exocrine glandular cells"                  
#>  [28] "pancreatic endocrine cells"                
#>  [29] "decidual cells"                            
#>  [30] "trophoblastic cells"                       
#>  [31] "myocytes"                                  
#>  [32] "fibroblasts"                               
#>  [33] "keratinocytes"                             
#>  [34] "Langerhans"                                
#>  [35] "melanocytes"                               
#>  [36] "epidermal cells"                           
#>  [37] "smooth muscle cells"                       
#>  [38] "peripheral nerve"                          
#>  [39] "cells in red pulp"                         
#>  [40] "cells in white pulp"                       
#>  [41] "cells in seminiferous ducts"               
#>  [42] "Leydig cells"                              
#>  [43] "urothelial cells"                          
#>  [44] "N/A"                                       
#>  [45] "follicle cells"                            
#>  [46] "chondrocytes"                              
#>  [47] "elongated or late spermatids"              
#>  [48] "pachytene spermatocytes"                   
#>  [49] "peritubular cells"                         
#>  [50] "preleptotene spermatocytes"                
#>  [51] "round or early spermatids"                 
#>  [52] "sertoli cells"                             
#>  [53] "spermatogonia cells"                       
#>  [54] "alveolar cells type I"                     
#>  [55] "alveolar cells type II"                    
#>  [56] "basal cells"                               
#>  [57] "ciliated cells (cell body)"                
#>  [58] "ciliated cells (cilia axoneme)"            
#>  [59] "ciliated cells (ciliary rootlets)"         
#>  [60] "ciliated cells (tip of cilia)"             
#>  [61] "goblet cells"                              
#>  [62] "cells in basal layer"                      
#>  [63] "cells in corneal layer"                    
#>  [64] "cells in spinous layer"                    
#>  [65] "extracellular matrix"                      
#>  [66] "fibrohistiocytic cells"                    
#>  [67] "hair follicles"                            
#>  [68] "langerhans cells"                          
#>  [69] "lymphocytes"                               
#>  [70] "vascular mural cells"                      
#>  [71] "Bergmann glia - cytoplasm/membrane"        
#>  [72] "Bergmann glia - nucleus"                   
#>  [73] "GLUC cells - cytoplasm/membrane"           
#>  [74] "GLUC cells - nucleus"                      
#>  [75] "granular cells - cytoplasm/membrane"       
#>  [76] "granular cells - nucleus"                  
#>  [77] "molecular layer - neuropil"                
#>  [78] "molecular layer cells - cytoplasm/membrane"
#>  [79] "molecular layer cells - nucleus"           
#>  [80] "processes in granular layer"               
#>  [81] "processes in molecular layer"              
#>  [82] "processes in white matter"                 
#>  [83] "Purkinje cells - cytoplasm/membrane"       
#>  [84] "Purkinje cells - dendrites"                
#>  [85] "Purkinje cells - nucleus"                  
#>  [86] "synaptic glomeruli - capsule"              
#>  [87] "synaptic glomeruli - core"                 
#>  [88] "white matter cells - cytoplasm/membrane"   
#>  [89] "white matter cells - nucleus"              
#>  [90] "endocrine cells"                           
#>  [91] "enterocytes"                               
#>  [92] "enterocytes - Microvilli"                  
#>  [93] "mucosal lymphoid cells"                    
#>  [94] "glands of Brunner"                         
#>  [95] "paneth cells"                              
#>  [96] "bowman's capsule"                          
#>  [97] "collecting ducts"                          
#>  [98] "distal tubules"                            
#>  [99] "proximal tubules (cell body)"              
#> [100] "proximal tubules (microvilli)"             
#> [101] "arrector pili muscle cells"                
#> [102] "non-ciliated cells"                        
#> [103] "cytotrophoblasts"                          
#> [104] "hofbauer cells"                            
#> [105] "syncytiotrophoblasts - cell body"          
#> [106] "syncytiotrophoblasts - microvilli"         
#> [107] "sebaceous glands"                          
#> [108] "eccrine glands"                            
#> [109] "neuronal projections"                      
#> [110] "synapses"                                  
#> [111] "ependymal cells"                           
#> [112] "ciliated cells"                            
#> [113] "ciliated epithelial cells"                 
#> [114] "nonciliated glandular epithelial cells"    
#> [115] "nonciliated luminal epithelial cells"      
#> [116] "stromal fibroblasts"                       
#> [117] "cells in cortex/medulla"                   
#> [118] "cells in cuticle"                          
#> [119] "cells in external root sheath"             
#> [120] "cells in internal root sheath"             
#> [121] "enterocytes - Gradient"                    
#> [122] "ganglion cells"                            
#> [123] "inner nuclear layer"                       
#> [124] "inner plexiform layer"                     
#> [125] "limiting membrane"                         
#> [126] "nerve fiber layer"                         
#> [127] "outer plexiform layer"                     
#> [128] "photoreceptor cells"                       
#> [129] "pigment epithelial cells"                  
#> [130] "lactating glandular cells"                 
#> [131] "cells in zona fasciculata"                 
#> [132] "cells in zona glomerulosa"                 
#> [133] "cells in zona reticularis"                 
#> [134] "medullary cells"                           
#> [135] "sebaceous cells"                           
#> [136] "secretory cells"                           
#> [137] "sweat ducts"                               
#> [138] "syncytiotrophoblasts"                      
#> [139] "granulosa cells"                           
#> [140] "oocytes"                                   
#> [141] "stroma"                                    
#> [142] "cortical cells"                            
#> [143] "cells in dentate nucleus"                  
#> [144] "corneal epithelial cells"                  
#> [145] "hyaloid membrane"                          
#> [146] "lens epithelial cells"                     
#> [147] "lens fiber cells"                          
#> [148] "cells in anterior"                         
#> [149] "cells in posterior"                        
#> [150] "extravillous trophoblasts"                 
#> [151] "ductal cells"                              
#> [152] "atretic follicle"                          
#>