Download datasets

Download the latest version of HPA datasets and import them in R. It is recommended to only download the datasets you need, as some of them may be very big.

hpaDownload(downloadList = "histology", version = "latest")

Arguments

downloadList
A vector or string indicate which datasets to download. Possible value:

'Normal tissue'

'Pathology'

'Subcellular location'

'RNA consensus tissue'

'RNA HPA tissue'

'RNA GTEx tissue'

'RNA FANTOM tissue'

'RNA single cell type'

'RNA single cell type tissue cluster'

'RNA GTEx brain region'

'RNA FANTOM brain region'

'RNA pig brain region'

'RNA pig brain subregion sample'

'RNA mouse brain region'

'RNA mouse brain subregion sample'

'RNA Allen mouse brain region'

'RNA HPA blood cell'

'RNA HPA blood cell sample'

'RNA Monaco blood cell'

'RNA Schmiedel blood cell'

'RNA HPA cell line'

'RNA TCGA cancer sample'

'RNA transcript tissue'

'RNA transcript cell line'

'RNA transcript pig brain'

'RNA transcript mouse brain'

'all': download everything

'histology': same as c('Normal tissue', 'Pathology', 'Subcellular location')

'rna tissue': same as c('RNA consensus tissue', 'RNA HPA tissue', 'RNA GTEx tissue', 'RNA FANTOM tissue')

'rna single cell type': same as c('RNA single cell type', 'RNA single cell type tissue cluster')

'rna brain region': same as c('RNA GTEx brain region', 'RNA FANTOM brain region', 'RNA pig brain region', 'RNA pig brain subregion sample', 'RNA mouse brain region', 'RNA mouse brain subregion sample', 'RNA Allen mouse brain region')

'rna single cell type': same as c('RNA HPA blood cell', 'RNA HPA blood cell sample', 'RNA Monaco blood cell', 'RNA Schmiedel blood cell')

'isoform': same as c('RNA transcript tissue', 'RNA transcript cell line', 'RNA transcript pig brain', 'RNA transcript mouse brain')

See https://www.proteinatlas.org/about/download for more information.

version
A string indicate which version to be downloaded. Possible value:

'latest': Download latest version. Require Internet connection.

'example' or 'built-in': Load the built-in histology dataset from 'HPAanalyze' ('hpa_histology_data'). Do not require internet connection.

Value

This function will return a list of tibbles corresponding to requested datasets.

See also

Examples

  downloadedData <- hpaDownload(downloadList='histology', version='example')

#> Only the followings are example/built-in datasets: 
#>  - Normal tissue 
#>  - Pathology 
#>  - Subcellular location 
#> Other datasets will not be loaded

  summary(downloadedData)

#>                      Length Class  Mode
#> normal_tissue         6     tbl_df list
#> pathology            11     tbl_df list
#> subcellular_location 14     tbl_df list

  
  downloadedDataRna <- hpaDownload(downloadList='rna tissue', version='latest')
  summary(downloadedDataRna)

#>                      Length Class  Mode
#> rna_tissue_consensus 4      tbl_df list
#> rna_tissue_hpa       6      tbl_df list
#> rna_tissue_gtex      6      tbl_df list
#> rna_tissue_fantom    6      tbl_df list