Download datasets

Download the latest version of HPA datasets and import them in R. It is recommended to only download the datasets you need, as some of them may be very big.

hpaDownload(downloadList = "histology", version = "latest")

Arguments

downloadList
A vector or string indicate which datasets to download. Possible value:

'Normal tissue'

'Pathology'

'Subcellular location'

'RNA tissue'

'RNA cell line'

'RNA transcript tissue'

'RNA transcript cell line'

'all': download everything

'histology': same as c('Normal tissue', 'Pathology', 'Subcellular location')

'rna': same as c('RNA tissue', 'RNA cell line')

'isoform': same as c('RNA transcript tissue', 'RNA transcript cell line')

See https://www.proteinatlas.org/about/download for more information.

version
A string indicate which version to be downloaded. Possible value:

'hpar': Load data from the 'hpar' package. Does not contain isoform data.

'latest': Download latest version. Require Internet connection.

'example': Load the example dataset from 'HPA analyze' ('hpa_downloaded_histology_v18'). Does not contain rna or isoform data.

'v?' with '?' is a integer: Download a specific version of the dataset. For example: 'v18' download version 18. Currently support version 17 and above. Require Internet connection.

Value

This function will return a list of tibbles corresponding to requested datasets.

See also

Examples

  downloadedData <- hpaDownload(downloadList='all', version='example')
  summary(downloadedData)

#>                      Length Class  Mode
#> normal_tissue        5      tbl_df list
#> pathology            7      tbl_df list
#> subcellular_location 8      tbl_df list