Subset downloaded data

hpaSubset() subsets data by gene name, tissue, cell type, cancer and/or cell line. The input is the list object generated by hpaDownload() or as the output of another hpaSubset(). Use hpaListParam() to see the list of available parameters for a specific list object. Will not work on isoform data.

hpaListParam() list available variables in downloaded data that can be used as parameters to subset the data via hpaSubset(). This function work with the data object generated by hpaDownload() or a previous call of hpaSubset().

hpaSubset(data = NULL, targetGene = NULL, targetTissue = NULL,
  targetCellType = NULL, targetCancer = NULL, targetCellLine = NULL)

hpaListParam(data = NULL)

Arguments

data
Input the list object generated by hpaDownload() or hpaSubset()

targetGene
Vector of strings of HGNC gene symbols. It will be used to subset every dataset in the list object. You can also mix HGNC gene symbols and ensemnl ids (start with ENSG) and they will be converted to HGNC gene symbols.

targetTissue
Vector of strings of normal tissues. Will be used to subset the normal_tissue and rna_tissue dataset.

targetCellType
Vector of strings of normal cell types. Will be used to subset the normal_tissue dataset.

targetCancer
Vector of strings of cancer types. Will be used to subset the pathology dataset.

targetCellLine
Vector of strings of cell lines. Will be used to subset the rna_cell_line dataset.

Value

hpaSubset will return a list of tibbles as the result of subsetting, depending on the input data.

The output of hpaListParam() is a list of vectors containing all subset parameter for the downloaded data.

See also

Examples

  downloadedData <- hpaDownload(downloadList='all', version='example')
  geneList <- c('TP53', 'EGFR')
  tissueList <- c('breast', 'cerebellum', 'skin 1')
  cancerList <- c('breast cancer', 'glioma', 'melanoma')

  subsetData <- hpaSubset(data=downloadedData,
                          targetGene=geneList,
                          targetTissue=tissueList,
                          targetCancer=cancerList)

  downloadedData <- hpaDownload(downloadList='all', version='example')
  params <- hpaListParam(data=downloadedData)
  params$normal_tissue

#>  [1] "adrenal gland"     "appendix"          "bone marrow"      
#>  [4] "breast"            "bronchus"          "caudate"          
#>  [7] "cerebellum"        "cerebral cortex"   "cervix, uterine"  
#> [10] "colon"             "duodenum"          "endometrium 1"    
#> [13] "endometrium 2"     "epididymis"        "esophagus"        
#> [16] "fallopian tube"    "gallbladder"       "heart muscle"     
#> [19] "hippocampus"       "kidney"            "liver"            
#> [22] "lung"              "lymph node"        "nasopharynx"      
#> [25] "oral mucosa"       "ovary"             "pancreas"         
#> [28] "parathyroid gland" "placenta"          "prostate"         
#> [31] "rectum"            "salivary gland"    "seminal vesicle"  
#> [34] "skeletal muscle"   "skin 1"            "skin 2"           
#> [37] "small intestine"   "smooth muscle"     "soft tissue 1"    
#> [40] "soft tissue 2"     "spleen"            "stomach 1"        
#> [43] "stomach 2"         "testis"            "thyroid gland"    
#> [46] "tonsil"            "urinary bladder"   "vagina"           
#> [49] "hypothalamus"      "hair"              "lactating breast" 
#> [52] "skin"              "retina"            "eye"              
#> [55] "thymus"            "pituitary gland"   "choroid plexus"   
#> [58] "sole of foot"