Extract tissue expression and download images

Extract tissue expression information and url to download images from imported xml document resulted from hpaXmlGet().

hpaXmlTissueExprSum(importedXml, downloadImg = FALSE)

Arguments

importedXml

Input an xml document object resulted from a hpaXmlGet() call.

downloadImg

Logical argument. The function will download all image from the extracted urls into the working folder.

Value

This function return a list consists of a summary string, which is a very brief description of the protein, and a tibble of 2 columns: tissue (name of tissue available) and imageUrl (link to download the perspective image)

See also

Examples

if (FALSE) {
  GCH1xml <- hpaXmlGet('ENSG00000131979')
  hpaXmlTissueExprSum(GCH1xml)
}